refactor wizard feature
trying to get the language of the test a little more high level.
BDD best practice is to stick as close as possible to client/domain language
and avoid technical details as much as possible.

These changes make the test more robust in the face of changes to
site structure and button labels (the step definitions have to
change, but the feature file itself should be able to remain
unchanged even if we split the wizard across multiple pages,
renamed buttons, etc.)