don't allow Next button to work if it is disabled
I think something changed in bootstrap at some point without us
noticing. The 'Next' button is disabled on most sections until all the
questions have been answered. That still works as far as the bootstrap
CSS. The button clearly appears disabled. However, it still triggers the
click event, so the user can still move forward without answering the
questions.

So, explicitly check the disabled state of the button when it's clicked
and don't advance if it is disabled.

(also change the section id of the one section that can always advance).