PMT #115795: Email Copy Updates
This commit implements a multipart email so that users will recieve an
html message, and plain text as a fall back.

The HTML email was needed to render a table of TIPS for users based on
their imput. It seemed impractical to try to get this layout to work in
pure plain text, so I though HTML would be best.

Because of the layout issues, I left out the table from the plain text
version.